refactor: colocate quest template manager tests
